9235: moved setPdfFormFields to useCaseHelpers to address PR comments
