+++ web-api/src/applicationContext.js
@@ -876,7 +876,6 @@ const {
 } = require('../../shared/src/business/useCases/trialSessions/runTrialSessionPlanningReportInteractor');
 const {
   sanitizePdfInteractor,
-  setPdfFields,
 } = require('../../shared/src/business/useCases/pdf/sanitizePdfInteractor');
 const {
   saveCalendarNoteInteractor,
@@ -965,6 +964,9 @@ const {
 const {
   setNoticesForCalendaredTrialSessionInteractor,
 } = require('../../shared/src/business/useCases/trialSessions/setNoticesForCalendaredTrialSessionInteractor');
+const {
+  setPdfFormFields,
+} = require('../../shared/src/business/useCaseHelper/pdf/setPdfFormFields');
 const {
   setPriorityOnAllWorkItems,
 } = require('../../shared/src/persistence/dynamo/workitems/setPriorityOnAllWorkItems');
@@ -1803,7 +1805,7 @@ module.exports = (appContextUser, logger = createLogger()) => {
         sendIrsSuperuserPetitionEmail,
         sendServedPartiesEmails,
         serveDocumentAndGetPaperServicePdf,
-        setPdfFields,
+        setPdfFormFields,
         updateAssociatedJudgeOnWorkItems,
         updateCaseAndAssociations,
         updateCaseAutomaticBlock,
